Reestructura de carpetas: views/, models/, viewmodels/, etc.
