--- .idea/deviceManager.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="DeviceTable">
-    <option name="columnSorters">
-      <list>
-        <ColumnSorterState>
-          <option name="column" value="API" />
-          <option name="order" value="DESCENDING" />
-        </ColumnSorterState>
-      </list>
-    </option>
-  </component>
-</project>
