--- .idea/misc.xml
@@ -1,5 +1,0 @@
-<project version="4">
-  <component name="ProjectRootManager">
-    <output url="file://$PROJECT_DIR$/out" />
-  </component>
-</project>
