+++ lib/main.dart
@@ -1,6 +1,6 @@
-import 'screens/login.dart';
+import 'views/login.dart';
 import 'package:flutter/material.dart';
-import '../strings/strings_utils.dart';
+import '../resources/strings_utils.dart';
 
 void main() => runApp(MyApp());
 
@@ -13,4 +13,5 @@ class MyApp extends StatelessWidget {
     );
   }
 }
+
